21 ext req case file (#22)
Note:
Even though the commands `edit` and `change_status` are marked as ToDo and coming soon:tm: I'm going to merge it. Will be a problem for future me (or someone else) to implement it. *(The checks etc work already, just the actual execution needs to be done)*


* add role_play.file_case (first code)

* add more participants

* add German translation

* add about command

* rename `file case` to `case file`

* rename `file_case` to `case_file`

forgot it in original commit

* add first code to add new case-files

* /case-file create now has accusation

* creation of case-files is not possible

* fix issue with "bot is thinking..."

last commit should have had "now" and not "not" inside the commit-message

* add file_case.permission and fix core:permission

* remove code-duplication (`db.commit()`)

* fix translation-files

* add DE as available language

* adapt code to new style on develop

* add descriptions for permissions

* add an overview to the about command

* add commit after creating a new file case

* add base for /edit

* add base for /change_status